+++ tests/index.ts
@@ -18,3 +18,4 @@ import "./08-join";
 import "./09-orderby";
 import "./10-groupbyFns";
 import "./11-orm";
+// import "./12-history";